Move many_threads_sequential to broken list, fix deps
The many_threads_sequential test is currently broken as we leak all
thread stacks and don't return the correct exit code on thread
creation failure, so move it to the broken list so runtests still
works.

This also fixes the dependencies - the target(s) referenced by the
package config must express a dependency on all binaries that the
config references as the packaging step depends on this label.  Since
the libc-tests config label is just "//libc-tests" (reasonable enough)
this target must depend on a group that depends on the executables
listed in the config.

Change-Id: Ib159f6c1e31816a1932fb8417382779682460e10